geometrydrill is the new utils
